00-lab-getting-started/requirements.txt
openai==1.78.1
azure.identity==1.16.1
python-dotenv==1.0.0